[fix] fix data type for price in cakes and flavour table

+++ database/migrations/2024_07_25_033644_create_flavours_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('flavours', function (Blueprint $table) {
             $table->id();
             $table->string('name', 255);
-            $table->integer('price');
+            $table->double('price', 8, 2);
             $table->timestamps();
         });
     }
